try to fix double

+++ Class.C
@@ -26,8 +26,8 @@ double Class::get_current_ratio() const
 {
 	if (teacher_vec_.size() != 0)
 	{
-		double return_value = child_vec_.size() / teacher_vec_.size();
-		return (return_value); /*This is the current ratio of children to teachers*/
+		double return_value = (double)child_vec_.size() / teacher_vec_.size();
+		return return_value; /*This is the current ratio of children to teachers*/
 	}
 	return 0;
 }
